requirements.txt
langchain
ipykernel
langchain-openai
python-dotenv
langchain-groq
langchain-community
bs4
openai
unstructured
pypdf
wikipedia
sentence-transformers
arxiv
pymupdf
langchain-text-splitters
chromadb
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain-ollama
transformers
duckduckgo-search
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
duckduckgo-search
pytube
duckdb
huggingface_hub
numexpr